refactor(test): change get all questions positions gRPC endpoint

+++ api-gateway/src/main/java/com/nitb/apigateway/dto/UserTest/response/ResultDetailsResponseDto.java
@@ -25,7 +25,7 @@ public class ResultDetailsResponseDto {
 
     private float accuracy;
 
-    private int secondsSpent;
+    private long secondsSpent;
 
     private List<ResultDetailResponseDto> details;
 }
